添加一组对比改进法

+++ parahomework1/para1/Sum/excess/main.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 #define maxN 10000000
 int a[maxN];
-int sum,sum1=0,sum2=0;
+long long sum,sum1=0,sum2=0;
 
 int main()
 {
